feat: remove marketingEnabled flag from Schedule & Details
The `marketingEnabled` field in the course settings API was previously driven
by the `ENABLE_MKTG_SITE` backend setting, which has been removed in
openedx/openedx-platform#38720. The API now always returns `true`.

- Remove `marketingEnabled` prop from `BasicSection`; always render the
  promotional banner (the `true` path)
- Extract the legacy enrollment card (LMS about-page link + "Invite your
  students" mailto button) into a standalone `CoursePromotionCard` component
  exported from `basic-section`, so operators can restore it via the
  `PageBannerSlot` if desired
- Extend `PageBannerSlot` to pass `lmsLinkForAboutPage`, `courseDisplayName`,
  and `platformName` as `pluginProps` so `CoursePromotionCard` has what it needs
  when used from a plugin
- Update `PageBannerSlot` README with instructions for restoring the enrollment
  card via the slot
