+++ src/data/api.ts
@@ -239,7 +239,6 @@ export interface CourseSettingsData {
   languageOptions: [string, string][];
   lmsLinkForAboutPage: string;
   licensingEnabled: boolean;
-  marketingEnabled: boolean;
   mfeProctoredExamSettingsUrl: string;
   platformName: string;
   possiblePreRequisiteCourses: {
