+++ src/schedule-and-details/__mocks__/courseSettings.js
@@ -25,7 +25,6 @@ module.exports = {
     ['uk', 'Ukrainian'],
   ],
   lmsLinkForAboutPage: 'http://localhost:18000/courses/course-v1:edX+M12+2T2023/about',
-  marketingEnabled: true,
   mfeProctoredExamSettingsUrl: '',
   possiblePreRequisiteCourses: [
     {
