+++ src/schedule-and-details/basic-section/BasicSection.test.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { IntlProvider } from '@edx/frontend-platform/i18n';
 
-import { INVITE_STUDENTS_LINK_ID } from './constants';
 import messages from './messages';
 import BasicSection from '.';
 
@@ -19,7 +18,6 @@ describe('<BasicSection />', () => {
     courseNumber: 'bar number',
     run: 'foo run',
     lmsLinkForAboutPage: 'link://to',
-    marketingEnabled: true,
     courseDisplayName: 'foo course',
     platformName: 'Your platform name here',
   };
@@ -40,41 +38,13 @@ describe('<BasicSection />', () => {
     expect(getByText(props.run)).toBeInTheDocument();
   });
 
-  it('shows the page banner if the marketingEnabled is true', () => {
-    const { getByText, queryAllByText } = render(<RootWrapper {...props} />);
+  it('shows the page banner', () => {
+    const { getByText } = render(<RootWrapper {...props} />);
     expect(
       getByText(`Promoting your course with ${props.platformName}`),
     ).toBeInTheDocument();
     expect(
       getByText(messages.basicBannerText.defaultMessage),
     ).toBeInTheDocument();
-    expect(queryAllByText('Course summary page').length).toBe(0);
-  });
-
-  it('shows the course promotion if the marketingEnabled is false', () => {
-    const initialProps = { ...props, marketingEnabled: false };
-    const { getByText, getByRole, queryAllByText } = render(
-      <RootWrapper {...initialProps} />,
-    );
-    const inviteButton = getByRole('button', {
-      name: messages.basicPromotionButton.defaultMessage,
-    });
-
-    expect(getByText(/Course Summary Page/i)).toBeInTheDocument();
-    expect(
-      getByText(/(for student enrollment and access)/i),
-    ).toBeInTheDocument();
-    expect(getByText(props.lmsLinkForAboutPage)).toBeInTheDocument();
-    expect(inviteButton).toBeInTheDocument();
-    expect(queryAllByText(`Promoting your course with ${props.platformName}`).length).toBe(0);
-  });
-
-  it('checks link link to invite', () => {
-    const initialProps = { ...props, marketingEnabled: false };
-    const { getByTestId } = render(<RootWrapper {...initialProps} />);
-    const inviteLink = getByTestId(INVITE_STUDENTS_LINK_ID);
-    expect(decodeURIComponent(inviteLink.href)).toEqual(
-      `mailto:${process.env.INVITE_STUDENTS_EMAIL_TO}?body=The course ${props.courseDisplayName}, provided by ${props.platformName}, is open for enrollment. Please navigate to this course at ${props.lmsLinkForAboutPage} to enroll.&subject=Enroll in ${props.courseDisplayName}.`,
-    );
   });
 });
