+++ src/schedule-and-details/basic-section/index.jsx
@@ -1,60 +1,25 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import { useIntl, FormattedMessage } from '@edx/frontend-platform/i18n';
-import {
-  Button,
-  Card,
-  MailtoLink,
-  Hyperlink,
-} from '@openedx/paragon';
+import { useIntl } from '@edx/frontend-platform/i18n';
+
 import PageBannerSlot from '@src/plugin-slots/PageBannerSlot';
-import { Email as EmailIcon } from '@openedx/paragon/icons';
 
 import SectionSubHeader from '../../generic/section-sub-header';
-import { INVITE_STUDENTS_LINK_ID } from './constants';
+import CoursePromotionCard from './CoursePromotionCard';
 import messages from './messages';
+
+export { CoursePromotionCard };
 
 const BasicSection = ({
   org,
   courseNumber,
   run,
   lmsLinkForAboutPage,
-  marketingEnabled,
   courseDisplayName,
   platformName,
 }) => {
   const intl = useIntl();
   const [showPageBanner, setShowPageBanner] = useState(true);
-  const emailSubject = intl.formatMessage(
-    {
-      id: 'course-authoring.schedule.basic.email.subject',
-      defaultMessage: 'Enroll in {courseDisplayName}.',
-    },
-    { courseDisplayName },
-  );
-
-  const emailBody = intl.formatMessage(
-    {
-      id: 'course-authoring.schedule.basic.email.body',
-      defaultMessage:
-        'The course {courseDisplayName}, provided by {platformName}, is open for enrollment. Please navigate to this course at {lmsLinkForAboutPage} to enroll.',
-    },
-    {
-      courseDisplayName,
-      platformName,
-      lmsLinkForAboutPage,
-    },
-  );
-
-  const promotionTitle = (
-    <FormattedMessage
-      id="course-authoring.schedule.basic.promotion.title"
-      defaultMessage="Course summary page {smallText}"
-      values={{
-        smallText: <small>(for student enrollment and access)</small>,
-      }}
-    />
-  );
 
   const courseBasicInfo = [
     {
@@ -85,45 +50,15 @@ const BasicSection = ({
     <PageBannerSlot
       show={showPageBanner}
       onDismiss={() => setShowPageBanner(false)}
+      lmsLinkForAboutPage={lmsLinkForAboutPage}
+      courseDisplayName={courseDisplayName}
+      platformName={platformName}
     >
       <h4 className="text-black">{intl.formatMessage(messages.basicBannerTitle, { platformName })}</h4>
       <span className="text text-gray-700 text-left">
         {intl.formatMessage(messages.basicBannerText)}
       </span>
     </PageBannerSlot>
-  );
-
-  const renderCoursePromotion = () => (
-    <Card>
-      <Card.Header
-        className="h4 px-3 text-gray-500"
-        title={promotionTitle}
-        size="sm"
-      />
-      <Card.Section className="px-3 py-1">
-        <Hyperlink
-          destination={lmsLinkForAboutPage}
-          className="lead info-500 small text-decoration-none"
-          target="_blank"
-          showLaunchIcon={false}
-        >
-          {lmsLinkForAboutPage}
-        </Hyperlink>
-      </Card.Section>
-      <Card.Divider />
-      <Card.Footer className="p-3 justify-content-start">
-        <MailtoLink
-          to={process.env.INVITE_STUDENTS_EMAIL_TO}
-          subject={emailSubject}
-          body={emailBody}
-          data-testid={INVITE_STUDENTS_LINK_ID}
-        >
-          <Button variant="outline-primary" iconBefore={EmailIcon} size="sm">
-            {intl.formatMessage(messages.basicPromotionButton)}
-          </Button>
-        </MailtoLink>
-      </Card.Footer>
-    </Card>
   );
 
   return (
@@ -135,7 +70,7 @@ const BasicSection = ({
       <ul className="basic-info-list">
         {courseBasicInfo.map(renderBasicInfo)}
       </ul>
-      {marketingEnabled ? renderPageBanner() : renderCoursePromotion()}
+      {renderPageBanner()}
     </section>
   );
 };
@@ -145,7 +80,6 @@ BasicSection.propTypes = {
   courseNumber: PropTypes.string.isRequired,
   run: PropTypes.string.isRequired,
   lmsLinkForAboutPage: PropTypes.string.isRequired,
-  marketingEnabled: PropTypes.bool.isRequired,
   courseDisplayName: PropTypes.string.isRequired,
   platformName: PropTypes.string.isRequired,
 };
